feat: 增加ALTER CASCADE 1. schedule新增前端刪除及管理會員選單
-----------------------------------------------
test:刪除行程主題要刪除全部有關連的資料

1.測試可連動刪除有關連的資料內容後，要修改schedule controller的內容

------------------------------------------------
memo:
1.補login畫面
2.補manager成員的畫面
3.把前端刪除api註解調

+++ IdentityJWT/Controllers/WeatherController.cs
@@ -27,7 +27,7 @@ namespace PinPinServer.Controllers
         /// <returns>回應格式為JSON，架構為一個Day的列表裡頭包含兩個天氣資訊，分別為上下午，資訊為WeatherDataDTO格式</returns>
         //GET:api/Weather
         [HttpGet]
-        public async Task<ActionResult> Get(double lat, double lon, string units)
+        public async Task<ActionResult> Get(decimal lat, decimal lon, string units)
         {
             int? userID = _getUserId.PinGetUserId(User).Value;
             if (userID == null || userID == 0) return BadRequest("Invalid user ID");
